requirements.txt
asgiref==3.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.5.0
whitenoise==6.9.0
